+++ CHANGELOG.md
@@ -1,10 +1,20 @@
 # Changelog
 
-## Alpha 0.0.2 (tba)
+## Beta 0.1 (16.04.2020)
+
+First public beta release.
+
+### Added
+
+- You can now bind horn
+- You can now look behind your left back
+- In Gifu when you try to look back, you'll lean automatically to the left
 
 ### Changes
 
-- Updated for Mod Loader version 1.1.6
+- Updated for Mod Loader version 1.1.7
+- Changed default look angle to 60
+- Renamed "Angle" to "Look Angle"
 
 ### Bug Fixes